Fix Makefile and freeze requirements
