+++ .flake8
@@ -1,2 +1,2 @@
 [flake8]
-max-line-length = 88
+max-line-length = 93
