+++ .pylintrc
@@ -3,6 +3,10 @@ disable=bad-continuation,
         missing-module-docstring,
         missing-function-docstring,
         missing-class-docstring,
+        too-many-instance-attributes,
+        line-too-long,
+        invalid-name,
+        logging-format-interpolation
 
 
 max-module-lines=1070
