+++ requirements.txt
@@ -1,0 +1,25 @@
+appdirs==1.4.3
+astroid==2.3.3
+attrs==19.3.0
+black==19.10b0
+certifi==2020.4.5.1
+chardet==3.0.4
+click==7.1.1
+colorama==0.4.3
+entrypoints==0.3
+flake8==3.7.9
+idna==2.9
+isort==4.3.21
+lazy-object-proxy==1.4.3
+mccabe==0.6.1
+pathspec==0.8.0
+pycodestyle==2.5.0
+pyflakes==2.1.1
+pylint==2.4.4
+regex==2020.4.4
+requests==2.23.0
+six==1.14.0
+toml==0.10.0
+typed-ast==1.4.1
+urllib3==1.25.9
+wrapt==1.11.2
